weather_week: treat empty samples better
Data saved by weather(1) can sometimes contain empty samples: samples
with a timestamp but no sampled values.  Plot these as gaps in the graph,
instead of the curves jumping to 0.0 degrees, 0.0 m/s and so on.

Could be done better (samples are included if they have at least one
non-zero value) but it's better than nothing.