weather_week: optimize SVG size a bit
Decrease the size a bit by using relative <path> movements, while
being careful not to introduce accumulating rounding errors.

Could do more here, by rendering 0.3 as .3, and 3.0 as 3.  Or by
scaling the view so that decimals can be skipped and we save all
those dots.

There's also gzipped SVG, but it's unclear to me how well supported
it is, and I couldn't get it to work with my web server and Firefox.