weather_week: render wind direction, as text
Plotting just the wind force (and max force) looks nice, but when
you look at it, you immediately want to know its direction, too.

The direction is printed as text, during periods of stability only.
You don't get to see the winds change in detail.

Also note that Trafikverket's data is of questionable quality in this
particular case.