weather_week: make wind direction output optional
If you want wind direction, you now have to ask for it with the
-w option.

It became gradually clear to me that I couldn't trust the input
weather data in this case.  Wind force seems to be captured ok
normally, but the direction is often just wrong.  It became obvious
last weekend in Falköping, when I noted a rather strong wind from W.
The graphs said SE.