weather: fix rain.amount
WeatherObservation contains Aggregated10minutes/Precipitation/
TotalWaterEquivalent, so we must scale that up to get mm/h,
which is what we always had in rain.amount (even though it was
undocumented by Trafikverket until recently).

For some reason, WeatherStation and WeatherObservation don't match
exactly even with this adjustment.  Perhaps WeatherStation does an
incorrect transformation or interpolates in an unknown way.