use Trafikinfo API version 2
Since 1.3 got shut down on 2022-10-26.

The change is trivial, boring and annoying:

    http://api.trafikinfo.trafikverket.se/v1.3/data.xml
    https://api.trafikinfo.trafikverket.se/v2/data.xml

plus you have to say schemaversion='2' in the query.

And it seems at some point over the years, TRV added
the GMT offset to the timestamps.

This commit doesn't implement https, so it's still useless.
Earlier API versions were available over plain HTTP, but
for no good reasons, v2 is not.